# Small performance optimization to string handling